cargo_build_script: add rule attr to take environment vars from a file
This patch adds a new cargo_build_script attribute called
"build_script_env_file".  This attribute is used to pass runtime
environment variables to the cargo build script via a text file.

This is similar in spirit to the rust_binary() rule "rustc_env_files"
attribute.

The format of the environment text file is:

- The variable name and value are defined on a single line as
  NAME=VALUE delimited by a newline `\n` character.

- If the VALUE contains embedded newlines, the embedded newlines are
  encoded as `%%%`.

Signed-off-by: Curt Brune <curt@enfabrica.net>