rust: add `use_rustc_linker` attribute to binary rules
This patch adds a new boolean attribute to the rust_XYZ() rules called
`use_rustc_linker`.  The default of this attribute is `False`.

When this attribute is `True`, the linker bundled with the `rustc`
distribution will be used instead of the linker from the cc_toolchain.
The built-in linker is $(RUST_SYSROOT)/lib/rustlib/<HOST_TRIPLE>/bin/rust-lld.

When this attribute is `False` (the default), the linker from the
cc_toolchain is used, which is typically GNU `gcc`.

Signed-off-by: Curt Brune <curt@enfabrica.net>