rust_analyzer: support other platforms
This patch adds a `platforms` CLI argument to the `gen_rust_project`
binary.  The argument is passed through to underlying `bazel build`
and `bazel aquery` commands that build the rust analzyer .json file.

The default value of this parameter is
`@bazel_tools//platforms:target_platform`, which is bazel's default
for the `--platforms` parameter.

This allows cross-platform users to specify for which platform the
specified targets should be analyzed.

Signed-off-by: Curt Brune <curt@enfabrica.net>